Выделил функции добавления контента на страницу
1. Функция addHTMLContent добавляет результат запроса (data) в элемент div#app и возвращает коллекцию кнопок. Данная функция вызывается в процессе преобразования возвращаемой ею коллекции кнопок в массив с последующим итерированием
2. Функция addBtnListeners добавляет обработчики события ('click') на каждую кнопку и определяет кнопку, по которой был произведен клик. Данная кнопка передается параметром во вложенную функцию addBtnContent
3. Функция addBtnContent принимает в качестве параметра кнопку, по которой был зарегистрирован клик функцией addBtnListeners, и, в зависимости от id кнопки, добавляет один из двух вариантов текстового контента
