+++ src/js/__tests__/bigUntestableCode.js
@@ -1,6 +1,8 @@
-import { getItemsRequest, toggleFavoriteRequest } from './requests';
+import getItemsRequest from './requests';
 import displayErrorElement from './displayError';
 import displayLoaderElement from './displayLoader';
+import addHTMLContent from './addHTMLContent';
+import addBtnListeners from './addBtnListeners';
 
 export default () => {
     displayErrorElement(false);
@@ -10,29 +12,11 @@ export default () => {
         .then(({ data }) => {
             if (data.result !== 'ok' || typeof data.html === 'undefined') {
                 displayErrorElement(true, 'Произошла ошибка, попробуйте ещё раз.');
-            } else {
-                const appElement = document.querySelector('#app');
-
-                appElement.innerHTML = data.html;
-                appElement.style.display = 'block';
-
-                Array.from(appElement.querySelector('button')).forEach((button) => {
-                    button.addEventListener('click', (e) => {
-                        e.preventDefault();
-
-                        e.currentTarget.textContent = 'Загрузка...';
-
-                        toggleFavoriteRequest(e.currentTarget.dataset.id)
-                            .then(({ data: buttonData }) => {
-                                if (buttonData.result === 'set') {
-                                    e.currentTarget.textContent = '🌝';
-                                } else {
-                                    e.currentTarget.textContent = '🌚';
-                                }
-                            });
-                    });
-                });
             }
+
+            Array.from(addHTMLContent(data)).forEach((button) => {
+                addBtnListeners(button);
+            });
         })
         .catch((e) => {
             displayErrorElement(true, e.message);
